+++ lib/Extension/Navigation/Navigator/Navigator.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Phpactor\Extension\Navigation\Navigator;
+
+interface Navigator
+{
+    public function destinationsFor(string $path): array;
+}