+++ lib/Extension/Navigation/Navigator/PathFinderNavigator.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Phpactor\Extension\Navigation\Navigator;
+
+use Phpactor\ClassFileConverter\Exception\NoMatchingSourceException;
+use Phpactor\ClassFileConverter\PathFinder;
+
+class PathFinderNavigator implements Navigator
+{
+    /**
+     * @var PathFinder
+     */
+    private $pathFinder;
+
+    public function __construct(PathFinder $pathFinder)
+    {
+        $this->pathFinder = $pathFinder;
+    }
+
+    public function destinationsFor(string $path): array
+    {
+        try {
+            return $this->pathFinder->destinationsFor($path);
+        } catch (NoMatchingSourceException $e) {
+            return [];
+        }
+    }
+}