+++ CHANGELOG.md
@@ -3,8 +3,10 @@ Changelog
 
 ## Develop
 
-Functions:
+Features:
 
+  - [Navigation] Reflection navigation: navigate to related classes (currently
+    supports parent class and interfaces).
   - [Completion] _Experimental_ class completion: complete use, new and
     extends. Class names inferred from file names.